backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
mysql-connector-python==8.0.26
pytest==6.2.5
pytest-cov==2.12.1
logging==0.4.9.6
python-json-logger==2.0.2
gunicorn==20.1.0  # For production deployment
